+++ altimeter/aws/auth/exceptions.py
