+++ bin/aws2json.py
@@ -41,7 +41,7 @@ from pathlib import Path
 import sys
 from typing import Dict, List, Set
 
-from altimeter.aws.access.v2.accessor import Accessor
+from altimeter.aws.auth import Accessor
 from altimeter.core.awslambda import get_required_lambda_env_var
 from altimeter.core.graph.graph_set import GraphSet
 from altimeter.core.log import Logger
